+++ src/date/index.js
@@ -1,0 +1,7 @@
+import { getPureDate } from "./getPureDate/index.js";
+import { isValidDate } from "./isValidDate/index.js";
+
+export {
+  isValidDate,
+  getPureDate
+};