+++ src/evt/onWindowLoad/index.js
@@ -8,7 +8,7 @@ import { getDocument, getWindow } from "ssr-window";
  * @see https://developer.mozilla.org/en-US/docs/Web/API/Window/load_event
  * @return {{ handler: Function, addListener: Function, removeListener: Function }}
  * @example
- * // How to detect when whole page has loaded?
+ * // How to detect when whole page (window) has loaded?
  * const callback = () => console.log("Page loaded");
  * onWindowLoad(callback);
  * */