+++ src/fn/index.js
@@ -1,5 +1,6 @@
 import { getCurryFn } from "./getCurryFn/index.js";
 import { getDebouncedFn } from "./getDebouncedFn/index.js";
+import { getThrottledFn } from "./getThrottledFn/index.js";
 import { isFnAsync } from "./isFnAsync/index.js";
 import { isFnClass } from "./isFnClass/index.js";
 import { wait } from "./wait/index.js";
@@ -10,4 +11,5 @@ export {
   isFnClass,
   isFnAsync,
   wait,
+  getThrottledFn,
 };